nimbusvault/auth-service/requirements.txt
fastapi
uvicorn
PyJWT[crypto]
passlib[bcrypt]
